+++ cogs/FunAndGames.py
@@ -12,17 +12,6 @@ class FunAndGames(commands.Cog):
     def __init__(self, bot):
         self.bot = bot
 
-
-    @commands.command(name='reload', hidden=True)
-    async def _reload(self, ctx, *, arg):
-        try:
-            self.bot.reload_extension(arg)
-        except commands.ExtensionError as e:
-            await ctx.send(f'{e.__class__.__name__}: {e}')
-        else:
-            await ctx.send('Reloaded {}'.format(arg))
-
-
     @commands.command(name='echo')
     async def echo(self, ctx, *args):
 	    await ctx.send(' '.join(args))
